Update Aggregation.swift
Edited for spelling and clarity. One comment that I'll paste in here without creating a review for it.
line 158
is a having statement a regular and known thing to your audience, the way that a Sort by statement would presumably be? if so, then this line should be sufficient. if not, provide more context/info here.