Update AttachmentSet.swift
Edited for clarity and grammar. You probably want to decide whether to use the term "AttachmentSet" or "attachment set" and then use it consistently. 

Please check lines 14, 37, 290, and 306 to make sure that I did not change the meaning of those lines. I am not going to create a review for this one, just a pull request.