Update Options.swift
Edited for clarity and grammar. No detailed comments except for this one, which applies to the whole document:
I changed "within" to "with". Please check all instances of this change and the change of param to parameter to make sure that the lines are still accurate.
If you mean that X is inside Y, then use "within" (Options object within the appURI parameter set.) 
If you mean that something contains or has something else, then use "with" (Options object with the appURI parameter set.)
These phrases were difficult to parse because set could be a verb (the appURI parameter was set to a value) or could be a noun (a set of parameters for appURI)